I've been experimenting with approaches to the Multiplicative Persistence problem, and in the name of experimentation I've committed various sins, such as single-letter variables, a lack of comments and docstrings, and not using version control. Now that the code is working and I feel I'm on the right track, I'm going to start using version control, and refactor it into something more legible. The old code is in the file persistence.py, which will be removed in a future version.
In the process, I'm going to add support for logging and multiprocessing, as well as writing unit tests.

Multiplicative Persistence is the number of iterations you can make multiplying the digits of a positive integer together, and repeating the operation on the
result, until you reach a single-digit number. It is conjectured that there are no numbers with a persistence greater than 11. This program will attempt to find
a counterexample.

Because I am doing this recreationally, I have not looked into the literature regarding this problem, and would be unsurprised to learn that the technique I am
employing has already been attempted. However, I intend to give it a try and to see what I come up with.

The idea is to find a number who's digits multiply together to form a number with a known multiplicative persistence of 11. The smallest number with a
persistence of 11, 277777788888899, is prime; therefore, there is no set of digits we can multiply together that results in this number. However, we can
transform this number is several ways:

* Rearrange the digits - as multiplication is communicative, the order of digits does not matter; * Factor the digits - The 9s can be replaced with 33, the 8s
with 222, etc; * Insert 1s anywhere in the number - as 1 is the multiplicative identity, we can use as many as we'd like.

Using these rules, the third especially, we can produce an infinite series of numbers with a multiplicative persistence of 11, not all of which all are prime.

We then need to ask if this number is "reachable", that is, could this number have possibly come from the product of the digits of another number? Any number
which results from this operation must be factorable by the single-digit primes (2, 3, 5, and 7), and by no other primes. Conversely, for any number that is
factorable by the single-digit primes, if we take the prime factors & concatenate them together, we will have a number, when we run 1 iteration of the
multiplicative persistence operation, will result in that number.

Therefore, if we are able to transform the number 27777778888889 into another number which is "reachable", we will have found a counterexample to the above
conjecture. This program will search for just such a number.

For more information on Multiplicative Persistence, see the following references:

http://mathworld.wolfram.com/MultiplicativePersistence.html
http://oeis.org/A003001
https://www.youtube.com/watch?v=Wim9WJeDTHQ
https://www.youtube.com/watch?v=E4mrC39sEOQ