Added tests for persistence.core.enumerateDigits().
All types of bad input to this function should be caught by
normalize()/acceptsNumber(), so only it's functionality is tested.
This does suggest that some of the tests in normalize() should be factored
out, perhaps into a set of tests for acceptsNumber(), but this is not currently
a high priority.