Added the function persistence.core.compositeCombinations() (formerly known as enumerateDigits()).
This function starts with the prime digits of a number and groups them together into composite
digits, in such a way that does not change it's multiplicative persistence. It computes each
possible combination, and yields them as a tuple of digits.

The approach is to use a queue containing both the combinations we've discovered so far and the
path we took to get there. Items are popped from the queue, and the function tries to rearrange
the digits into each of the composite single-digit numbers. If it observes that a combination is
possible, it adds this composite digit to the current path to calculate the new path, and checks
to see if this is a path it has already reached some other way. Only if it is a novel path is it
yielded & put into the queue. This way, each path is returned only once, and the queue does not
grow indefinitely.

I am confident that this works & reaches every combination, but I can't quite see how to prove
it mathematically yet.