Removed tests for type & value errors from test_digits_to_int.py.
I removed theses tests because checking for these errors is the responsibility of
the @acceptsNumber wrapper, and because this function isn't likely to recieve
pathological input anyway (given the non-userfacing nature of this codebase).

As this is a performance-driven application, without security considerations and
modest reliability considerations, duplicating error checking is imprudent.