Implemented digitsToInt().
This function converts digits, as represented by a tuple of integers, into an integer.
This function is used by reachable(). All other functions need to work with numbers
in their tuple representation, to preserve their base 10 properties, but reachable()
need to work with them as native Python integers, in order to test them with the
modulus (%) operator.