Updated requirements.txt to include hypothesis.
Hypothesis is a framework for testing invariants. It works with py.test,
so no refactoring of old tests is needed. This will allow me to express
powerful ideas, such as, "for any number that is reachable, it's factorization
should have a persistence of exactly 1 more than the original number's
persistence."

Hypothesis will then employ fuzzing to search for values that break these
invariants.

Documentation: https://hypothesis.readthedocs.io/en/latest/