Added property testing to test.test_composit_combinations.py.
Tested the invariant that the testcase itself is inside the output of
persistence.core.compositeCombinations(). If there is some possibility
that is neglected, then there should be some number for which this
invariant doesn't hold, and visa versa - if there is no number for
which this invariant doesn't hold, there are no possibilities that
are neglected.

I had to create a new Hypothesis strategy because of the behavior of
persistence.core.normalize(). 0s would cause ValueErrors, and 1s would
be stripped out, creating false positives.

I also removed the comments explaining why the type testing was removed,
fixed a typo in anohther comment, and made a note that the remaining
type checking should be removed. This sort of thing should be implemented
using mypy type annotations.