Factored idiom for taking product of iterable into utility function
In working on tests for persistence.core.previous, I found myself using the idiom
reduce(lambda a,b: a*b, iterable) quite a bit. I also used this in
persistence.core.persistence. It is, afterall, the central operation to my problem.

I researched and found that this was added to the standard library in Python 3.8.
However, I am not yet on 3.8. So I wrote a version of it, with the intention to
remove it when distros catch up to 3.8.